requirements.txt
python-dotenv
langchain-groq
pandas
streamlit